etherscan-plugin/requirements.txt
Flask==2.1.1
Flask-Cors==3.1.1
requests==2.27.1
python-dotenv==0.19.2
PyYAML==6.0